+++ src/main/java/fr/insa/lyon/pld/agile/model/Section.java
@@ -1,9 +1,5 @@
 package fr.insa.lyon.pld.agile.model;
 
-/**
- *
- * @author scheah
- */
 public class Section {
     public static final double SPEED = 15.*1000./3600.; //15 km/h => 15.*1000./3600./s
     
@@ -66,6 +62,6 @@ public class Section {
     }
     
     public static long computeDuration(double length) {
-        return (long) Math.round(length/SPEED);
+        return Math.round(length/SPEED);
     }
 }
